requirements.txt
asgiref==3.4.1
asn1crypto==1.5.1
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.2.23
django-filter==21.1
djangorestframework==3.14.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.6
importlib-metadata==4.8.3
incremental==16.10.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
Markdown==3.3.7
mysqlclient==2.1.1
netifaces==0.10.4
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
PyGObject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pytz==2023.3.post1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sqlparse==0.4.4
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing_extensions==4.1.1
ufw==0.36
urllib3==1.22
zipp==3.6.0
zope.interface==4.3.2
